requirements/prod.txt
Django==2.2.5
djangorestframework==3.10.3
django-extensions==2.2.1

pyyaml==5.1.2
uritemplate==3.0.0
